+++ CHANGELOG.md
@@ -9,9 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * [BREAKING] [MAJOR] Split the main module into two packages:
-    * azkustodata - contains querying, management and commands.
+    * azkustodata - contains querying, management APIs.
     * azkustoingest - contains ingestion in all its forms.
-* [BREAKING] Minimum Go version is now 1.20.
 ### Fixed
 - Modified `once.go` to reset `sync.Once` instance when an error occurs